requirements.txt
langchain==0.1.13
langchain_core==0.1.36
langchain_openai==0.1.1
openai==1.14.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0
streamlit==1.32.2
